Integrating custom clients
