+++ src/Benchmarks.ClientJob/ClientJob.cs
@@ -21,9 +21,11 @@ namespace Benchmarks.ClientJob
             Threads = clientJob.Threads;
             Connections = clientJob.Connections;
             Duration = clientJob.Duration;
+            ClientProperties = new Dictionary<string, string>(clientJob.ClientProperties);
+            ClientName = clientJob.ClientName;
             ScriptName = clientJob.ScriptName;
             PipelineDepth = clientJob.PipelineDepth;
-            Headers = clientJob.Headers;
+            Headers = new Dictionary<string, string>(clientJob.Headers);
             ServerBenchmarkUri = clientJob.ServerBenchmarkUri;
             Query = clientJob.Query;
             State = clientJob.State;
@@ -36,17 +38,18 @@ namespace Benchmarks.ClientJob
 
         public int Id { get; set; }
 
+        public string ClientName { get; set; }
+
         public int Threads { get; set; } = 32;
 
         public int Connections { get; set; } = 256;
 
         public int Duration { get; set; } = 15;
         public int Warmup { get; set; } = 15;
-
+        public Dictionary<string, string> ClientProperties { get; set; } = new Dictionary<string, string>();
+        // TODO: Remove when Clients have been upgraded to the new protocol
         public string ScriptName { get; set; }
-
         public int PipelineDepth { get; set; }
-
         public Dictionary<string, string> Headers { get; set; } = new Dictionary<string, string>();
 
         public string ServerBenchmarkUri { get; set; }
@@ -70,7 +73,7 @@ namespace Benchmarks.ClientJob
         public string Method { get; set; } = "GET";
         public bool SkipStartupLatencies { get; set; }
 
-        // Latency of  first request
+        // Latency of first request
         public TimeSpan LatencyFirstRequest { get; set; }
 
         // Latency with a single connection
