+++ src/Benchmarks/benchmarks.html.json
@@ -1,6 +1,7 @@
 {
   "Default": {
 
+    "ClientName": "wrk",
     "PresetHeaders": "Html",
 
     "Source": {
