+++ src/Benchmarks/benchmarks.json.json
@@ -1,6 +1,7 @@
 {
   "Default": {
 
+    "ClientName": "wrk",
     "PresetHeaders": "Json",
 
     "Source": {
