+++ src/Benchmarks/benchmarks.multiquery.json
@@ -1,6 +1,7 @@
 {
   "Default": {
 
+    "ClientName": "wrk",
     "PresetHeaders": "Json",
 
     "Query": "?queries=20",
