+++ src/Benchmarks/benchmarks.plaintext.json
@@ -1,10 +1,16 @@
 {
   "Default": {
 
-    "PresetHeaders": "Plaintext",
+    "ClientName": "wrk",
+    "ClientProperties": {
+      "ScriptName": "pipeline",
+      "PipelineDepth": 16
+    },
+
     "ScriptName": "pipeline",
     "PipelineDepth": 16,
 
+    "PresetHeaders": "Plaintext",
     "Source": {
       "Repository": "https://github.com/aspnet/benchmarks.git",
       "BranchOrCommit": "dev",
@@ -16,6 +22,10 @@
   },
   "PlaintextNonPipelined": {
     "Path": "/plaintext",
+    "ClientProperties": {
+      "ScriptName": "",
+      "PipelineDepth": 0
+    },
     "ScriptName": "",
     "PipelineDepth": 0
   },
