+++ src/BenchmarksDriver/Program.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Data.SqlClient;
 using System.IO;
 using System.Linq;
 using System.Net;
@@ -13,10 +12,10 @@ using System.Threading.Tasks;
 using System.Web;
 using Benchmarks.ClientJob;
 using Benchmarks.ServerJob;
+using BenchmarksWorkers;
 using McMaster.Extensions.CommandLineUtils;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
-using OperatingSystem = Benchmarks.ServerJob.OperatingSystem;
 
 namespace BenchmarksDriver
 {
@@ -44,6 +43,8 @@ namespace BenchmarksDriver
             // Driver Options
             var clientOption = app.Option("-c|--client",
                 "URL of benchmark client", CommandOptionType.SingleValue);
+            var clientNameOption = app.Option("--clientName",
+                "Name of client to use for testing, e.g. 'wrk'.", CommandOptionType.SingleValue);
             var serverOption = app.Option("-s|--server",
                 "URL of benchmark server", CommandOptionType.SingleValue);
             var sqlConnectionStringOption = app.Option("-q|--sql",
@@ -136,10 +137,8 @@ namespace BenchmarksDriver
                 "Default set of HTTP headers added to request (None, Plaintext, Json, Html). Default is Html.", CommandOptionType.SingleValue);
             var methodOption = app.Option("--method",
                 "HTTP method of the request. Default is GET.", CommandOptionType.SingleValue);
-            var scriptNameOption = app.Option("--script",
-                "Name of the script used by wrk.", CommandOptionType.SingleValue);
-            var pipelineDepthOption = app.Option("--pipelineDepth",
-                "Depth of pipeline used by client.", CommandOptionType.SingleValue);
+            var clientProperties = app.Option("--properties",
+                "Key value pairs of properties specific to the client running. e.g., Threads=16;PipelineDepth=16", CommandOptionType.SingleValue);
             var pathOption = app.Option(
                 "--path",
                 "Relative URL where the client should send requests.",
@@ -475,6 +474,17 @@ namespace BenchmarksDriver
                 mergedClientJob.Merge(job);
                 _clientJob = mergedClientJob.ToObject<ClientJob>();
 
+                if (clientNameOption.HasValue())
+                {
+                    _clientJob.ClientName = clientNameOption.Value();
+                }
+
+                if (String.IsNullOrEmpty(_clientJob.ClientName))
+                {
+                    Console.WriteLine($"The client name is missing.");
+                    return 11;
+                }
+
                 // Override default ClientJob settings if options are set
                 if (connectionsOption.HasValue())
                 {
@@ -492,19 +502,16 @@ namespace BenchmarksDriver
                 {
                     _clientJob.Warmup = int.Parse(warmupOption.Value());
                 }
-                if (pipelineDepthOption.HasValue())
+                if (clientProperties.HasValue())
                 {
-                    _clientJob.PipelineDepth = int.Parse(pipelineDepthOption.Value());
-
-                    if (_clientJob.PipelineDepth > 0)
+                    var properties = clientProperties.Value().Split(';');
+                    foreach (var kvp in properties)
                     {
-                        _clientJob.ScriptName = "pipeline";
+                        var values = kvp.Split('=');
+                        _clientJob.ClientProperties[values[0]] = values[1];
                     }
                 }
-                if (scriptNameOption.HasValue())
-                {
-                    _clientJob.ScriptName = scriptNameOption.Value();
-                }
+
                 if (methodOption.HasValue())
                 {
                     _clientJob.Method = methodOption.Value();
@@ -622,10 +629,16 @@ namespace BenchmarksDriver
 
             var results = new List<Statistics>();
             ClientJob clientJob = null;
+            IResultsSerializer serializer = null;
 
             if (!string.IsNullOrWhiteSpace(sqlConnectionString))
             {
-                await EnsureTableCreated(sqlConnectionString);
+                serializer = WorkerFactory.CreateResultSerializer(clientJob);
+
+                if (serializer != null)
+                {
+                    await serializer.InitializeDatabaseAsync(sqlConnectionString, _tableName);
+                }
             }
 
             var content = JsonConvert.SerializeObject(serverJob);
@@ -972,195 +985,20 @@ namespace BenchmarksDriver
                                 Log($"Socket Errors:               {average.SocketErrors}");
                                 Log($"Bad Responses:               {average.BadResponses}");
 
-                                if (!string.IsNullOrWhiteSpace(sqlConnectionString))
+                                if (serializer != null)
                                 {
                                     Log("Writing results to SQL...");
 
-                                    await WriteJobsToSql(
+                                    await serializer.WriteJobResultsToSqlAsync(
                                         serverJob: serverJob,
                                         clientJob: clientJob,
                                         connectionString: sqlConnectionString,
+                                        tableName: _tableName,
                                         path: serverJob.Path,
                                         session: session,
                                         description: description,
-                                        dimension: "RequestsPerSecond",
-                                        value: average.RequestsPerSecond);
-
-                                    if (average.StartupMain != -1 && spanLoop == 0)
-                                    {
-                                        await WriteJobsToSql(
-                                        serverJob: serverJob,
-                                        clientJob: clientJob,
-                                        connectionString: sqlConnectionString,
-                                        path: serverJob.Path,
-                                        session: session,
-                                        description: description,
-                                        dimension: "Startup Main (ms)",
-                                        value: average.StartupMain);
-                                    }
-
-                                    if (average.FirstRequest != -1 && spanLoop == 0)
-                                    {
-                                        await WriteJobsToSql(
-                                        serverJob: serverJob,
-                                        clientJob: clientJob,
-                                        connectionString: sqlConnectionString,
-                                        path: serverJob.Path,
-                                        session: session,
-                                        description: description,
-                                        dimension: "First Request (ms)",
-                                        value: average.FirstRequest);
-                                    }
-
-                                    await WriteJobsToSql(
-                                        serverJob: serverJob,
-                                        clientJob: clientJob,
-                                        connectionString: sqlConnectionString,
-                                        path: serverJob.Path,
-                                        session: session,
-                                        description: description,
-                                        dimension: "WorkingSet (MB)",
-                                        value: average.WorkingSet);
-
-                                    await WriteJobsToSql(
-                                        serverJob: serverJob,
-                                        clientJob: clientJob,
-                                        connectionString: sqlConnectionString,
-                                        path: serverJob.Path,
-                                        session: session,
-                                        description: description,
-                                        dimension: "CPU",
-                                        value: average.Cpu);
-
-                                    if (average.Latency != -1 && spanLoop == 0)
-                                    {
-                                        await WriteJobsToSql(
-                                        serverJob: serverJob,
-                                        clientJob: clientJob,
-                                        connectionString: sqlConnectionString,
-                                        session: session,
-                                        description: description,
-                                        path: serverJob.Path,
-                                        dimension: "Latency (ms)",
-                                        value: average.Latency);
-                                    }
-
-                                    if (average.LatencyAverage != -1)
-                                    {
-                                        await WriteJobsToSql(
-                                            serverJob: serverJob,
-                                            clientJob: clientJob,
-                                            connectionString: sqlConnectionString,
-                                            path: serverJob.Path,
-                                            session: session,
-                                            description: description,
-                                            dimension: "LatencyAverage (ms)",
-                                            value: average.LatencyAverage);
-                                    }
-
-                                    if (average.Latency50Percentile != -1)
-                                    {
-                                        await WriteJobsToSql(
-                                        serverJob: serverJob,
-                                        clientJob: clientJob,
-                                        connectionString: sqlConnectionString,
-                                        path: serverJob.Path,
-                                        session: session,
-                                        description: description,
-                                        dimension: "Latency50Percentile (ms)",
-                                        value: average.Latency50Percentile);
-                                    }
-
-                                    if (average.Latency75Percentile != -1)
-                                    {
-                                        await WriteJobsToSql(
-                                        serverJob: serverJob,
-                                        clientJob: clientJob,
-                                        connectionString: sqlConnectionString,
-                                        path: serverJob.Path,
-                                        session: session,
-                                        description: description,
-                                        dimension: "Latency75Percentile (ms)",
-                                        value: average.Latency75Percentile);
-                                    }
-
-                                    if (average.Latency90Percentile != -1)
-                                    {
-                                        await WriteJobsToSql(
-                                        serverJob: serverJob,
-                                        clientJob: clientJob,
-                                        connectionString: sqlConnectionString,
-                                        path: serverJob.Path,
-                                        session: session,
-                                        description: description,
-                                        dimension: "Latency90Percentile (ms)",
-                                        value: average.Latency90Percentile);
-                                    }
-
-                                    if (average.Latency99Percentile != -1)
-                                    {
-                                        await WriteJobsToSql(
-                                        serverJob: serverJob,
-                                        clientJob: clientJob,
-                                        connectionString: sqlConnectionString,
-                                        path: serverJob.Path,
-                                        session: session,
-                                        description: description,
-                                        dimension: "Latency99Percentile (ms)",
-                                        value: average.Latency99Percentile);
-                                    }
-
-                                    if (average.SocketErrors != -1)
-                                    {
-                                        await WriteJobsToSql(
-                                        serverJob: serverJob,
-                                        clientJob: clientJob,
-                                        connectionString: sqlConnectionString,
-                                        path: serverJob.Path,
-                                        session: session,
-                                        description: description,
-                                        dimension: "SocketErrors",
-                                        value: average.SocketErrors);
-                                    }
-
-                                    if (average.BadResponses != -1)
-                                    {
-                                        await WriteJobsToSql(
-                                        serverJob: serverJob,
-                                        clientJob: clientJob,
-                                        connectionString: sqlConnectionString,
-                                        path: serverJob.Path,
-                                        session: session,
-                                        description: description,
-                                        dimension: "BadResponses",
-                                        value: average.BadResponses);
-                                    }
-
-                                    if (average.TotalRequests != -1)
-                                    {
-                                        await WriteJobsToSql(
-                                        serverJob: serverJob,
-                                        clientJob: clientJob,
-                                        connectionString: sqlConnectionString,
-                                        path: serverJob.Path,
-                                        session: session,
-                                        description: description,
-                                        dimension: "TotalRequests",
-                                        value: average.TotalRequests);
-                                    }
-
-                                    if (average.Duration != -1)
-                                    {
-                                        await WriteJobsToSql(
-                                        serverJob: serverJob,
-                                        clientJob: clientJob,
-                                        connectionString: sqlConnectionString,
-                                        path: serverJob.Path,
-                                        session: session,
-                                        description: description,
-                                        dimension: "Duration (ms)",
-                                        value: average.Duration);
-                                    }
+                                        statistics: average,
+                                        longRunning: span > TimeSpan.Zero);
                                 }
                             }
                         }
@@ -1412,241 +1250,6 @@ namespace BenchmarksDriver
         {
             var uri = serverJobUri + "/invoke?path=" + HttpUtility.UrlEncode(path);
             Console.WriteLine(await _httpClient.GetStringAsync(uri));
-        }
-
-        private static Task WriteJobsToSql(ServerJob serverJob, ClientJob clientJob, string connectionString, string path, string session, string description, string dimension, double value)
-        {
-            return RetryOnExceptionAsync(5, () =>
-                 WriteResultsToSql(
-                        connectionString: connectionString,
-                        scenario: serverJob.Scenario,
-                        session: session,
-                        description: description,
-                        aspnetCoreVersion: serverJob.AspNetCoreVersion,
-                        runtimeVersion: serverJob.RuntimeVersion,
-                        hardware: serverJob.Hardware.Value,
-                        hardwareVersion: serverJob.HardwareVersion,
-                        operatingSystem: serverJob.OperatingSystem.Value,
-                        scheme: serverJob.Scheme,
-                        sources: serverJob.ReferenceSources,
-                        connectionFilter: serverJob.ConnectionFilter,
-                        webHost: serverJob.WebHost,
-                        kestrelThreadCount: serverJob.KestrelThreadCount,
-                        clientThreads: clientJob.Threads,
-                        connections: clientJob.Connections,
-                        duration: clientJob.Duration,
-                        pipelineDepth: clientJob.PipelineDepth,
-                        path: path,
-                        method: clientJob.Method,
-                        headers: clientJob.Headers,
-                        dimension: dimension,
-                        value: value,
-                        runtimeStore: serverJob.UseRuntimeStore)
-            , 5000);
-        }
-
-        private static async Task EnsureTableCreated(string connectionString)
-        {
-            string createCmd =
-                @"
-                IF OBJECT_ID(N'dbo." + _tableName + @"', N'U') IS NULL
-                BEGIN
-                    CREATE TABLE [dbo].[" + _tableName + @"](
-                        [Id] [int] IDENTITY(1,1) NOT NULL PRIMARY KEY,
-                        [Excluded] [bit] DEFAULT 0,
-                        [DateTime] [datetimeoffset](7) NOT NULL,
-                        [Session] [nvarchar](200) NOT NULL,
-                        [Description] [nvarchar](200),
-                        [AspNetCoreVersion] [nvarchar](50) NOT NULL,
-                        [RuntimeVersion] [nvarchar](50) NOT NULL,
-                        [Scenario] [nvarchar](50) NOT NULL,
-                        [Hardware] [nvarchar](50) NOT NULL,
-                        [HardwareVersion] [nvarchar](128) NOT NULL,
-                        [OperatingSystem] [nvarchar](50) NOT NULL,
-                        [Framework] [nvarchar](50) NOT NULL,
-                        [RuntimeStore] [bit] NOT NULL,
-                        [Scheme] [nvarchar](50) NOT NULL,
-                        [Sources] [nvarchar](50) NULL,
-                        [ConnectionFilter] [nvarchar](50) NULL,
-                        [WebHost] [nvarchar](50) NOT NULL,
-                        [KestrelThreadCount] [int] NULL,
-                        [ClientThreads] [int] NOT NULL,
-                        [Connections] [int] NOT NULL,
-                        [Duration] [int] NOT NULL,
-                        [PipelineDepth] [int] NULL,
-                        [Path] [nvarchar](200) NULL,
-                        [Method] [nvarchar](50) NOT NULL,
-                        [Headers] [nvarchar](max) NULL,
-                        [Dimension] [nvarchar](50) NOT NULL,
-                        [Value] [float] NOT NULL
-                    )
-                END
-                ";
-
-            using (var connection = new SqlConnection(connectionString))
-            {
-                await connection.OpenAsync();
-
-                using (var command = new SqlCommand(createCmd, connection))
-                {
-                    await command.ExecuteNonQueryAsync();
-                }
-            }
-
-        }
-
-        private static async Task WriteResultsToSql(
-            string connectionString,
-            string session,
-            string description,
-            string aspnetCoreVersion,
-            string runtimeVersion,
-            string scenario,
-            Hardware hardware,
-            string hardwareVersion,
-            OperatingSystem operatingSystem,
-            Scheme scheme,
-            IEnumerable<Source> sources,
-            string connectionFilter,
-            WebHost webHost,
-            int? kestrelThreadCount,
-            int clientThreads,
-            int connections,
-            int duration,
-            int? pipelineDepth,
-            string path,
-            string method,
-            IDictionary<string, string> headers,
-            string dimension,
-            double value,
-            bool runtimeStore)
-        {
-            
-            string insertCmd =
-                @"
-                INSERT INTO [dbo].[" + _tableName + @"]
-                           ([DateTime]
-                           ,[Session]
-                           ,[Description]
-                           ,[AspNetCoreVersion]
-                           ,[RuntimeVersion]
-                           ,[Scenario]
-                           ,[Hardware]
-                           ,[HardwareVersion]
-                           ,[OperatingSystem]
-                           ,[Framework]
-                           ,[RuntimeStore]
-                           ,[Scheme]
-                           ,[Sources]
-                           ,[ConnectionFilter]
-                           ,[WebHost]
-                           ,[KestrelThreadCount]
-                           ,[ClientThreads]
-                           ,[Connections]
-                           ,[Duration]
-                           ,[PipelineDepth]
-                           ,[Path]
-                           ,[Method]
-                           ,[Headers]
-                           ,[Dimension]
-                           ,[Value])
-                     VALUES
-                           (@DateTime
-                           ,@Session
-                           ,@Description
-                           ,@AspNetCoreVersion
-                           ,@RuntimeVersion
-                           ,@Scenario
-                           ,@Hardware
-                           ,@HardwareVersion
-                           ,@OperatingSystem
-                           ,@Framework
-                           ,@RuntimeStore
-                           ,@Scheme
-                           ,@Sources
-                           ,@ConnectionFilter
-                           ,@WebHost
-                           ,@KestrelThreadCount
-                           ,@ClientThreads
-                           ,@Connections
-                           ,@Duration
-                           ,@PipelineDepth
-                           ,@Path
-                           ,@Method
-                           ,@Headers
-                           ,@Dimension
-                           ,@Value)
-                ";
-
-            using (var connection = new SqlConnection(connectionString))
-            {
-                await connection.OpenAsync();
-
-                using (var command = new SqlCommand(insertCmd, connection))
-                {
-                    var p = command.Parameters;
-                    p.AddWithValue("@DateTime", DateTimeOffset.UtcNow);
-                    p.AddWithValue("@Session", session);
-                    p.AddWithValue("@Description", description);
-                    p.AddWithValue("@AspNetCoreVersion", aspnetCoreVersion);
-                    p.AddWithValue("@RuntimeVersion", aspnetCoreVersion);
-                    p.AddWithValue("@Scenario", scenario.ToString());
-                    p.AddWithValue("@Hardware", hardware.ToString());
-                    p.AddWithValue("@HardwareVersion", hardwareVersion);
-                    p.AddWithValue("@OperatingSystem", operatingSystem.ToString());
-                    p.AddWithValue("@Framework", "Core");
-                    p.AddWithValue("@RuntimeStore", runtimeStore);
-                    p.AddWithValue("@Scheme", scheme.ToString().ToLowerInvariant());
-                    p.AddWithValue("@Sources", sources.Any() ? (object)ConvertToSqlString(sources) : DBNull.Value);
-                    p.AddWithValue("@ConnectionFilter",
-                        string.IsNullOrEmpty(connectionFilter) ? (object)DBNull.Value : connectionFilter);
-                    p.AddWithValue("@WebHost", webHost.ToString());
-                    p.AddWithValue("@KestrelThreadCount", (object)kestrelThreadCount ?? DBNull.Value);
-                    p.AddWithValue("@ClientThreads", clientThreads);
-                    p.AddWithValue("@Connections", connections);
-                    p.AddWithValue("@Duration", duration);
-                    p.AddWithValue("@PipelineDepth", (object)pipelineDepth ?? DBNull.Value);
-                    p.AddWithValue("@Path", string.IsNullOrEmpty(path) ? (object)DBNull.Value : path);
-                    p.AddWithValue("@Method", method.ToString().ToUpperInvariant());
-                    p.AddWithValue("@Headers", headers.Any() ? JsonConvert.SerializeObject(headers) : (object)DBNull.Value);
-                    p.AddWithValue("@Dimension", dimension);
-                    p.AddWithValue("@Value", value);
-
-                    await command.ExecuteNonQueryAsync();
-                }
-            }
-        }
-
-        private static string ConvertToSqlString(IEnumerable<Source> sources)
-        {
-            return string.Join(",", sources.Select(s => ConvertToSqlString(s)));
-        }
-
-        private static string ConvertToSqlString(Source source)
-        {
-            const string aspnetPrefix = "https://github.com/aspnet/";
-            const string gitSuffix = ".git";
-
-            var shortRepository = source.Repository;
-
-            if (shortRepository.StartsWith(aspnetPrefix))
-            {
-                shortRepository = shortRepository.Substring(aspnetPrefix.Length);
-            }
-
-            if (shortRepository.EndsWith(gitSuffix))
-            {
-                shortRepository = shortRepository.Substring(0, shortRepository.Length - gitSuffix.Length);
-            }
-
-            if (string.IsNullOrEmpty(source.BranchOrCommit))
-            {
-                return shortRepository;
-            }
-            else
-            {
-                return shortRepository + "@" + source.BranchOrCommit;
-            }
         }
 
         private static void QuietLog(string message)
