+++ src/BenchmarksWorkers/IWorker.cs
@@ -1,0 +1,16 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using System.Threading.Tasks;
+
+namespace BenchmarksWorkers
+{
+    public interface IWorker : IDisposable
+    {
+        string JobLogText { get; set; }
+
+        Task StartAsync();
+        Task StopAsync();
+    }
+}
